+++ pic/input.c
@@ -307,9 +307,9 @@ int nextchar(void)
 				FATAL("argfp underflow");
 			popsrc();
 			goto loop;
-		} else if (c == '$' && isdigit(*srcp->sp)) {
+		} else if (c == '$' && isdigit((int)*srcp->sp)) {
 			int n = 0;
-			while (isdigit(*srcp->sp))
+			while (isdigit((int)*srcp->sp))
 				n = 10 * n + *srcp->sp++ - '0';
 			if (n > 0 && n <= MAXARGS)
 				pushsrc(String, argfp->argstk[n-1]);
